+++ src/templates/react-app/react/base/src/App.mustache.js
@@ -3,17 +3,17 @@ import { Home } from './components/Home';
 {{#models}}
 import { {{modelNamePascal}}Router } from './components/{{modelNamePascal}}/{{modelNamePascal}}Router';
 {{/models}}
-import { Header } from "./components/App/Header";
+import { AppHeader } from "./components/App/AppHeader";
 import './App.scss';
 
 function App() {
   return (
     <Router>
       <div className="App">
-        <Header />
+        <AppHeader />
         <Switch>
         {{#models}}
-          <Route path="/{{modelNamePascal}}">
+          <Route path={["/{{modelNamePascal}}", "/{{pluralModelNamePascal}}"]}>
             <{{modelNamePascal}}Router />
           </Route>
           {{/models}}