+++ src/templates/react-app/react/base/src/components/Home.mustache.js
@@ -1,17 +1,20 @@
 import { Container, List, ListItem, ListItemText } from '@material-ui/core';
 import { Link } from 'react-router-dom';
+import { Header } from "./common/Header";
 
 export function Home() {
   return (
     <Container maxWidth="sm" className="Container Home">
-      <h1>Welcome to {{ appName }}!</h1>
-      <List component="nav">
-      {{#models}}
-        <ListItem button component={Link} to="/{{modelNamePascal}}">
-          <ListItemText primary="{{modelName}}" />
-        </ListItem>
-      {{/models}}
-      </List>
+       <Header title="Welcome to {{ appName }}!"/>
+       <div className="containerContent">
+        <List component="nav">
+        {{#models}}
+          <ListItem button component={Link} to="/{{pluralModelNamePascal}}">
+            <ListItemText primary="{{pluralModelName}}" />
+          </ListItem>
+        {{/models}}
+        </List>
+      </div>
     </Container>
   );
 }